# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Its a REST API Documentation for school app backend",
    "version": "1.0.0",
    "title": "SONET School App Documentation"
  },
  "host": "",
  "basePath": "/",
...
  "schemes": ["http"],
  "components": {
    "schemas": {
...
      "respObj": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string",
            "default": "OK"
          },
          "resultCode": {
            "type": "integer",
            "default": 200
          },
          "message": {
            "type": "string",
            "default": "Query Success"
          },
          "ValidationErrors": {
            "type": "any",
            "default": ""
          },
          "resultTotal": {
            "type": "integer",
            "default": 0
          },
          "totalRows": {
            "type": "integer",
            "default": 0
          }
        }
      },
      "getAClass": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "integer",
            "default": 1
          },
          "className": {
            "type": "string",
            "default": "Class-I"
          },
          "schoolId": {
            "type": "integer",
            "default": 1
          }
        }
      },
      "getAClassesWithoutSchool": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "integer",
            "default": 1
          },
          "className": {
            "type": "string",
            "default": "Class-I"
          },
          "school": {
            "type": "object",
            "properties": {
              "schoolName": {
                "type": "string",
                "default": "Talent High School"
              }
            }
          },
          "sections": {
            "$ref": "#/components/schemas/manageSections/getAllSectionsOfAClass/properties/data"
          }
        }
      },
      "listOfClasses": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/getAClass"
            }
          }
        }
      },
      "listOfClassesWithOutSchool": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/getAClassesWithoutSchool"
            }
          }
        }
      },
...
      "editAClass": {
        "type": "object",
        "properties": {
          "className": {
            "type": "string",
            "required": false
          },
          "status": {
            "type": "boolean",
            "returied": false
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/classes/create": {
      "post": {
        "tags": ["Manage Classes"],
        "summary": "Create classes",
        "description": "Create a new class under a school, school info will be taken from the backend",
        "operationId": "newClassCreation",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "className",
            "description": "Enter the new class name",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/getAClass"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Invalid credentials"
          },
          "200": {
            "description": "Login Success"
          }
        }
      }
    },
    "/api/classes/update/{classId}": {
      "patch": {
        "tags": ["Manage Classes"],
        "summary": "Update a class",
        "description": "Update a class",
        "operationId": "editAClass",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "classBody",
            "description": "",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/editAClass"
            }
          },
          {
            "name": "classId",
            "in": "path",
            "required": "true",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Class updated successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/listOfClasses"
                }
              ]
            }
          }
        }
      }
    },
    "/api/classes/{offset}/{limit}": {
      "get": {
        "tags": ["Manage Classes"],
        "summary": "List of classes",
        "description": "Show list of classes",
        "operationId": "ListOfClasses",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "offset",
            "in": "path",
            "required": false,
            "default": 0,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "path",
            "required": false,
            "default": 10,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Class updated successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/listOfClassesWithOutSchool"
                }
              ]
            }
          }
        }
      }
    },
    "/api/sections/": {
      "get": {
        "tags": ["Manage Sections"],
        "summary": "List of sections",
        "description": "Show list of sections of a defailtClass in a school",
        "operationId": "ListOfSections",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Listed sections successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSections/getAllSectionsOfAClass"
                }
              ]
            }
          }
        }
      }
    },
...
    "/api/sections/create": {
      "post": {
        "tags": ["Manage Sections"],
        "summary": "Create section/s",
        "description": "Create a new section/s under a class",
        "operationId": "createSection",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "SectionBody",
            "description": "Section requrest body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageSections/createSections"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Listed sections successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSections/getAllSectionsOfAClass"
                }
              ]
            }
          }
        }
      }
    },
    "/api/sections/update": {
      "patch": {
        "tags": ["Manage Sections"],
        "summary": "Edit section/s",
        "description": "Edit  section/s under a class",
        "operationId": "updateSection",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "SectionBody",
            "description": "Section requrest body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageSections/editSections"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Listed sections successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageSections/getAllSectionsOfAClass"
                }
              ]
            }
          }
        }
      }
    },
    "/api/students/": {
      "get": {
        "tags": ["Manage Students"],
        "summary": "List of students of a class",
        "description": "Show list of students of a class",
        "operationId": "studentsOfAClass",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "classId",
            "in": "query",
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Students of a class has been listed successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageStudents/getStudentsOfClass"
                }
              ]
            }
          }
        }
      }
    },
    "/api/students/create": {
      "post": {
        "tags": ["Manage Students"],
        "summary": "Add students",
        "description": "Add new students to a class",
        "operationId": "createStudents",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "RequestBodyy",
            "description": "Create students request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageStudents/createAStudents"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "New students has been created successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageStudents/getStudentsOfClass"
                }
              ]
            }
          }
        }
      }
    },
    "/api/students/update": {
      "patch": {
        "tags": ["Manage Students"],
        "summary": "Update students",
        "description": "Update students to a class",
        "operationId": "updateStudents",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/tokens/x-access-token"
          },
          {
            "$ref": "#/tokens/refreshToken"
          },
          {
            "name": "RequestBodyy",
            "description": "Create students request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/manageStudents/UpdateStudents"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "UnAuthorized user"
          },
          "200": {
            "description": "Students has been updated successfully",
            "content": "application/json",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/respObj"
                },
                {
                  "$ref": "#/components/schemas/manageStudents/getStudentsOfClass"
                }
              ]
            }
          }
        }
      }
    },
...
  },
...
}